Reorganise out of page.rs
